backend/lumen_agents/requirements.txt
httpx>=0.28.0
google-genai>=1.0.0
